pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glue-yaml-processor"
version = "1.0.0"
description = "A Python library for processing YAML configurations in AWS Glue Python shell jobs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["aws", "glue", "yaml", "etl", "mysql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "boto3>=1.26.0",
    "PyMySQL>=1.0.0",
    "PyYAML>=6.0",
    "aws-glue-libs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/glue-yaml-processor"
Repository = "https://github.com/yourusername/glue-yaml-processor"
Documentation = "https://github.com/yourusername/glue-yaml-processor#readme"
Issues = "https://github.com/yourusername/glue-yaml-processor/issues"

[project.scripts]
glue-yaml-processor = "glue_yaml_processor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["glue_yaml_processor"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true